requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
django-cors-headers==3.7.0

djangorestframework-simplejwt-5.3.1 pyjwt-2.8.0
django-tinymce-3.7.1
Pillow-10.2.0

<<<<<<< HEAD
=======
################################
>>>>>>> 749970f3ea87b628f1a409a0234452924fd0221f
